+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -625,7 +625,7 @@ type AmazonRdsForSQLServerSource struct {
 	SourceRetryWait interface{} `json:"sourceRetryWait,omitempty"`
 
 	// Value and type setting for stored procedure parameters. Example: "{Parameter1: {value: "1", type: "int"}}".
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 }
 
 // GetCopySource implements the CopySourceClassification interface for type AmazonRdsForSQLServerSource.
@@ -5134,7 +5134,7 @@ type AzureSQLSink struct {
 	SinkRetryWait interface{} `json:"sinkRetryWait,omitempty"`
 
 	// SQL stored procedure parameters.
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 
 	// The stored procedure parameter name of the table type. Type: string (or Expression with resultType string).
 	StoredProcedureTableTypeParameterName interface{} `json:"storedProcedureTableTypeParameterName,omitempty"`
@@ -5215,7 +5215,7 @@ type AzureSQLSource struct {
 	SourceRetryWait interface{} `json:"sourceRetryWait,omitempty"`
 
 	// Value and type setting for stored procedure parameters. Example: "{Parameter1: {value: "1", type: "int"}}".
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 }
 
 // GetCopySource implements the CopySourceClassification interface for type AzureSQLSource.
@@ -5554,6 +5554,11 @@ type AzureSynapseArtifactsLinkedServiceTypeProperties struct {
 	// Required to specify MSI, if using system assigned managed identity as authentication method. Type: string (or Expression
 	// with resultType string).
 	Authentication interface{} `json:"authentication,omitempty"`
+
+	// The resource ID of the Synapse workspace. The format should be: /subscriptions/{subscriptionID}/resourceGroups/{resourceGroup}/providers/Microsoft.Synapse/workspaces/{workspaceName}.
+	// Type: string (or
+	// Expression with resultType string).
+	WorkspaceResourceID interface{} `json:"workspaceResourceId,omitempty"`
 }
 
 // AzureTableDataset - The Azure Table storage dataset.
@@ -6388,27 +6393,6 @@ type ChainingTriggerTypeProperties struct {
 
 	// REQUIRED; Run Dimension property that needs to be emitted by upstream pipelines.
 	RunDimension *string `json:"runDimension,omitempty"`
-}
-
-// CloudError - The object that defines the structure of an Azure Data Factory error response.
-type CloudError struct {
-	// REQUIRED; Error data
-	Error *CloudErrorBody `json:"error,omitempty"`
-}
-
-// CloudErrorBody - The object that defines the structure of an Azure Data Factory error.
-type CloudErrorBody struct {
-	// REQUIRED; Error code.
-	Code *string `json:"code,omitempty"`
-
-	// REQUIRED; Error message.
-	Message *string `json:"message,omitempty"`
-
-	// Array with additional error details.
-	Details []*CloudError `json:"details,omitempty"`
-
-	// Property name/path in request associated with error.
-	Target *string `json:"target,omitempty"`
 }
 
 // CmdkeySetup - The custom setup of running cmdkey commands.
@@ -22767,7 +22751,7 @@ type SQLMISink struct {
 	SinkRetryWait interface{} `json:"sinkRetryWait,omitempty"`
 
 	// SQL stored procedure parameters.
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 
 	// The stored procedure parameter name of the table type. Type: string (or Expression with resultType string).
 	StoredProcedureTableTypeParameterName interface{} `json:"storedProcedureTableTypeParameterName,omitempty"`
@@ -22848,7 +22832,7 @@ type SQLMISource struct {
 	SourceRetryWait interface{} `json:"sourceRetryWait,omitempty"`
 
 	// Value and type setting for stored procedure parameters. Example: "{Parameter1: {value: "1", type: "int"}}".
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 }
 
 // GetCopySource implements the CopySourceClassification interface for type SQLMISource.
@@ -22983,7 +22967,7 @@ type SQLServerSink struct {
 	SinkRetryWait interface{} `json:"sinkRetryWait,omitempty"`
 
 	// SQL stored procedure parameters.
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 
 	// The stored procedure parameter name of the table type. Type: string (or Expression with resultType string).
 	StoredProcedureTableTypeParameterName interface{} `json:"storedProcedureTableTypeParameterName,omitempty"`
@@ -23064,7 +23048,7 @@ type SQLServerSource struct {
 	SourceRetryWait interface{} `json:"sourceRetryWait,omitempty"`
 
 	// Value and type setting for stored procedure parameters. Example: "{Parameter1: {value: "1", type: "int"}}".
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 }
 
 // GetCopySource implements the CopySourceClassification interface for type SQLServerSource.
@@ -23252,7 +23236,7 @@ type SQLSink struct {
 	SinkRetryWait interface{} `json:"sinkRetryWait,omitempty"`
 
 	// SQL stored procedure parameters.
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 
 	// The stored procedure parameter name of the table type. Type: string (or Expression with resultType string).
 	StoredProcedureTableTypeParameterName interface{} `json:"storedProcedureTableTypeParameterName,omitempty"`
@@ -23335,7 +23319,7 @@ type SQLSource struct {
 	SourceRetryWait interface{} `json:"sourceRetryWait,omitempty"`
 
 	// Value and type setting for stored procedure parameters. Example: "{Parameter1: {value: "1", type: "int"}}".
-	StoredProcedureParameters map[string]*StoredProcedureParameter `json:"storedProcedureParameters,omitempty"`
+	StoredProcedureParameters interface{} `json:"storedProcedureParameters,omitempty"`
 }
 
 // GetCopySource implements the CopySourceClassification interface for type SQLSource.
@@ -28184,12 +28168,20 @@ type SynapseSparkJobActivityTypeProperties struct {
 	// Expression with resultType string).
 	File interface{} `json:"file,omitempty"`
 
-	// Additional files used for reference in the main definition file, which will override the 'files' of the spark job definition
-	// you provide.
+	// (Deprecated. Please use pythonCodeReference and filesV2) Additional files used for reference in the main definition file,
+	// which will override the 'files' of the spark job definition you provide.
 	Files []interface{} `json:"files,omitempty"`
+
+	// Additional files used for reference in the main definition file, which will override the 'jars' and 'files' of the spark
+	// job definition you provide.
+	FilesV2 []interface{} `json:"filesV2,omitempty"`
 
 	// Number of executors to launch for this job, which will override the 'numExecutors' of the spark job definition you provide.
 	NumExecutors *int32 `json:"numExecutors,omitempty"`
+
+	// Additional python code files used for reference in the main definition file, which will override the 'pyFiles' of the spark
+	// job definition you provide.
+	PythonCodeReference []interface{} `json:"pythonCodeReference,omitempty"`
 
 	// The name of the big data pool which will be used to execute the spark batch job, which will override the 'targetBigDataPool'
 	// of the spark job definition you provide.
@@ -28254,8 +28246,8 @@ func (s *SynapseSparkJobDefinitionActivity) GetExecutionActivity() *ExecutionAct
 
 // SynapseSparkJobReference - Synapse spark job reference type.
 type SynapseSparkJobReference struct {
-	// REQUIRED; Reference spark job name.
-	ReferenceName *string `json:"referenceName,omitempty"`
+	// REQUIRED; Reference spark job name. Expression with resultType string.
+	ReferenceName interface{} `json:"referenceName,omitempty"`
 
 	// REQUIRED; Synapse spark job reference type.
 	Type *SparkJobReferenceType `json:"type,omitempty"`